L5

+++ src/app/contact/contact-list-item/contact-list-item.component.ts
@@ -2,6 +2,8 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Contact} from '../contact';
 import {ContactService} from '../services/contact.service';
 import {Router} from '@angular/router';
+import {MatDialog, MatSnackBar} from '@angular/material';
+import {ConfirmDialogComponent} from '../../ui/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-contact-list-item',
@@ -13,7 +15,10 @@ export class ContactListItemComponent implements OnInit {
   @Input() contact: Contact;
   @Output() contactSelect: EventEmitter<any>;
 
-  constructor(private router: Router, private contactService: ContactService) {
+  constructor(private router: Router,
+              private contactService: ContactService,
+              private snackBar: MatSnackBar,
+              private dialog: MatDialog) {
     this.contactSelect = new EventEmitter();
   }
 
@@ -26,7 +31,25 @@ export class ContactListItemComponent implements OnInit {
   }
 
   deleteItem() {
-    this.contactService.deleteContact(this.contact.id);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent);
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        console.log('Removing contact');
+        this.contactService.deleteContact(this.contact.id);
+
+        this.snackBar.open('Contact removed',
+          this.contact.firstName + ' ' + this.contact.lastName,
+          {
+            duration: 3000,
+            verticalPosition: 'top',
+            horizontalPosition: 'center'
+          });
+
+        this.contactSelect.emit(null);
+      } else {
+        console.log('Removing cancelled');
+      }
+    });
   }
 
   editItem() {
